requirements.txt
Flask==2.2.5
python-pptx==0.6.21
requests==2.31.0
Werkzeug>=2.2.5
